+++ neo/framework/async/MsgChannel.h
@@ -81,7 +81,7 @@ private:
 	byte			ReadByte( void );
 	void			WriteShort( int s );
 	int				ReadShort( void );
-	void			WriteLong( int l );
+	void			WriteInt( int l );
 	int				ReadLong( void );
 	void			WriteData( const byte *data, const int size );
 	void			ReadData( byte *data, const int size );
